Move keys to datamanager and fix potential circular import issues
